Reorg negotiate tests
